seeing photos on the galery page

+++ backend/src/models/albums/repositories/photos.repository.ts
@@ -1,9 +1,12 @@
 import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
+import { extname } from "path";
 import { PaginationOptions } from "src/helpers/pagination";
 import { Member } from "src/models/members/entities/member.entity";
+import { User } from "src/models/users/entities/user.entity";
 import { Repository } from "typeorm";
 import { Photo } from "../entities/photo.entity";
+import { PhotosFilesService } from "../services/photos-files.service";
 
 export interface GetPhotosOptions extends PaginationOptions {
 	album?: number;
@@ -11,16 +14,21 @@ export interface GetPhotosOptions extends PaginationOptions {
 
 @Injectable()
 export class PhotosRepository {
-	constructor(@InjectRepository(Photo) private repository: Repository<Photo>) {}
+	constructor(
+		@InjectRepository(Photo) private repository: Repository<Photo>,
+		private photosFiles: PhotosFilesService,
+	) {}
 
 	getPhotos(options: GetPhotosOptions = {}) {
-		const q = this.repository
-			.createQueryBuilder("photos")
-			.select(["photos.id", "photos.name", "photos.status"])
-			.take(options.limit || 25)
-			.skip(options.offset || 0);
+		const q = this.repository.createQueryBuilder("photos").orderBy("photos.timestamp", "ASC");
 
 		if (options.album) q.where("photos.album_id = :album", { album: options.album });
+
+		if (options.offset) q.skip(options.offset);
+
+		// return the whole album for the gallery; cap only the unfiltered global list
+		if (options.limit) q.take(options.limit);
+		else if (!options.album) q.take(50);
 
 		return q.getMany();
 	}
@@ -41,10 +49,29 @@ export class PhotosRepository {
 		return this.repository.findOneBy({ id });
 	}
 
-	async createPhoto(albumId: number, photo: Omit<Photo, "id">, file: Express.Multer.File) {
-		// TODO: save file
+	async createPhoto(albumId: number, file: Express.Multer.File, uploadedById: User["id"] | null) {
+		const ext = extname(file.originalname);
+		const metadata = await this.photosFiles.extractMetadata(file.buffer);
 
-		return this.repository.save({ photo, albumId });
+		const photo = await this.repository.save({
+			albumId,
+			uploadedById,
+			name: file.originalname,
+			timestamp: metadata.timestamp,
+			width: metadata.width,
+			height: metadata.height,
+			bg: metadata.bg,
+		});
+
+		try {
+			await this.photosFiles.savePhotoFiles(albumId, photo.id, ext, file.buffer);
+		} catch (err) {
+			await this.repository.delete(photo.id);
+			await this.photosFiles.deletePhotoFiles(albumId, photo.id, ext);
+			throw err;
+		}
+
+		return photo;
 	}
 
 	async updatePhoto(id: Photo["id"], photo: Partial<Photo>) {
@@ -52,8 +79,11 @@ export class PhotosRepository {
 	}
 
 	async deletePhoto(id: Photo["id"]) {
-		//TODO: delete files
-		return await this.repository.delete(id);
+		const photo = await this.repository.findOneBy({ id });
+		if (!photo) return;
+
+		await this.repository.delete(id);
+		await this.photosFiles.deletePhotoFiles(photo.albumId, photo.id, extname(photo.name));
 	}
 
 	async deletePhotosByAlbum(albumId: Photo["albumId"]) {
